refactor(dashboard): replace root barrel and wildcards with role entrypoints
Delete the dashboard-core root index barrel and drop the root and all
wildcard exports-map entries, leaving exactly the four role entrypoints
(runtime, state, selectors, widgets) as the package surface. The nine
per-widget subpath consumers consolidate onto the widgets entrypoint.

importBoundaries gains an assertion that Station production reaches
dashboard-core only through the role entrypoints, alongside the frozen
runtime-composition inventory. The station link script probes a role
entrypoint for dashboard-core now that no root dist entry exists.
