+++ packages/dashboard-core/src/entrypoints/widgets.ts
@@ -8,16 +8,10 @@
 
 export { resolveTopRowWidgets } from "../widgets/snapshotWidgets.js";
 export type {
-  SnapshotWidgetKind,
-  TimeWidgetRuntime,
   TopRowWidgetRuntimeDeps,
   TopRowWidgetView,
   TuiConfig,
   TuiIslandConfig,
   TuiWidgetConfig,
-  WeatherClient,
-  WeatherCurrentConditions,
-  WeatherTemperatureUnit,
 } from "../widgets/types.js";
-export type { TopRowWidgetHookRuntime } from "../widgets/useTopRowWidgets.js";
-export { createUseTopRowWidgets, refreshWeatherWidget } from "../widgets/useTopRowWidgets.js";
+export { createUseTopRowWidgets } from "../widgets/useTopRowWidgets.js";
