--- packages/dashboard-core/src/index.ts
@@ -1,54 +1,0 @@
-export * from "./components/BottomSheetFrame/layout.js";
-export * from "./components/Dashboard/content.js";
-export * from "./components/Dashboard/filterConditionPanel.js";
-export * from "./components/Dashboard/footer.js";
-export * from "./components/Dashboard/layout.js";
-export * from "./components/Dashboard/tableHeader.js";
-export * from "./components/EditableTextInput/editing.js";
-export * from "./components/HelpOverlay/helpPanel.js";
-export * from "./components/NewSessionBottomSheet/content.js";
-export * from "./components/NewSessionBottomSheet/layout.js";
-export * from "./components/ProjectSettingsPanel/layout.js";
-export * from "./components/TextMatch/segments.js";
-export * from "./components/ToastOverlay/content.js";
-export * from "./components/ToastOverlay/layout.js";
-export * from "./components/WidgetSettingsPanel/content.js";
-export * from "./components/WorktreeRow/layout.js";
-export * from "./components/WorktreeRow/rowInput.js";
-export * from "./flows/addProject/actions.js";
-export * from "./flows/addProject/flow.js";
-export * from "./flows/addProject/input.js";
-export * from "./flows/addProject/rows.js";
-export * from "./flows/addProject/state.js";
-export * from "./flows/addProject/types.js";
-export * from "./flows/newSession.js";
-export * from "./flows/stepWizard.js";
-export * from "./selectors/dashboardFilterConditions.js";
-export * from "./selectors/dashboardPersistentFilter.js";
-export * from "./selectors/dashboardSessionRows.js";
-export * from "./selectors/dashboardViewport.js";
-export * from "./selectors/fleetSummary.js";
-export * from "./selectors/selectors.js";
-export * from "./services/errors/errors.js";
-export * from "./services/folderService.js";
-export * from "./services/types.js";
-export * from "./state/capabilities/activation.js";
-export * from "./state/capabilities/dismissal.js";
-export * from "./state/capabilities/execution.js";
-export * from "./state/capabilities/managedSessions.js";
-export * from "./state/capabilities/shell.js";
-export * from "./state/commandBuilders.js";
-export * from "./state/dashboardScroll.js";
-export * from "./state/index.js";
-export * from "./state/keymap.js";
-export * from "./state/localRows.js";
-export * from "./state/screens/addProjectScreen.js";
-export * from "./state/screens/fork.js";
-export * from "./state/screens/projectDefaultAgent.js";
-export * from "./state/screens/projectSettings.js";
-export * from "./state/screens/quickSession.js";
-export * from "./state/screens/removeWorktree.js";
-export * from "./state/screens/sessionRows.js";
-export * from "./state/screens/widgetSettings.js";
-export * from "./state/selection/addProject.js";
-export * from "./state/selection/registry.js";
