+++ station/src/station/store/dashboardRuntime.ts
@@ -5,7 +5,7 @@
 // so onDismiss records that capability without owning the effect.
 import { createDashboardRuntime } from "@station/dashboard-core/runtime";
 import type { DashboardCapabilities, DashboardRuntime, TuiFolderService } from "@station/dashboard-core/runtime";
-import type { TuiWidgetConfig } from "@station/dashboard-core/widgets/types";
+import type { TuiWidgetConfig } from "@station/dashboard-core/widgets";
 import type { StationClient } from "../../sources/types.js";
 
 export type StationDashboardRuntime = DashboardRuntime & {
