+++ station/src/station/view/DashboardFrameTitle.tsx
@@ -4,8 +4,8 @@ import stringWidth from "string-width";
 import type { DashboardStateSource } from "@station/dashboard-core/runtime";
 import { headerStrip, observerHeaderStatusForConnection, selectFleetSummary } from "@station/dashboard-core/selectors";
 import { tuiScreenBehavior } from "@station/dashboard-core/state";
-import { resolveTopRowWidgets } from "@station/dashboard-core/widgets/snapshotWidgets";
-import type { TopRowWidgetView } from "@station/dashboard-core/widgets/types";
+import { resolveTopRowWidgets } from "@station/dashboard-core/widgets";
+import type { TopRowWidgetView } from "@station/dashboard-core/widgets";
 import { toOpenTuiColor, toOpenTuiOpaqueColor, useStationTheme } from "../../theme/index.js";
 import {
   stationMouseProps,
