+++ station/src/station/widgets/useTopRowWidgets.ts
@@ -1,4 +1,4 @@
-import { createUseTopRowWidgets } from "@station/dashboard-core/widgets/useTopRowWidgets";
+import { createUseTopRowWidgets } from "@station/dashboard-core/widgets";
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 
 export const useTopRowWidgets = createUseTopRowWidgets({
